#IoT-13 Fix conversion formula (RSSI to distance)

+++ mesh_network_server/main/lib/bluetooth_mesh.c
@@ -189,12 +189,12 @@ void update_ibeacon_state(uint8_t *uuid, uint16_t major, uint16_t minor, int rss
     //    S - measured signal value (RSSI) in dBm
     //    N - environmental factor (usually value between 2 and 4)
     int meausuredPower = -74;
-    float envFactor = 5.3;
+    double envFactor = 5.3;
 
     memcpy(_ibeacon_model_state.uuid, uuid, 16);
     _ibeacon_model_state.major = major;
     _ibeacon_model_state.minor = minor;
     _ibeacon_model_state.rssi = rssi;
-    _ibeacon_model_state.distance = pow(10,((float )(meausuredPower - rssi) / 10 * envFactor));
+    _ibeacon_model_state.distance = pow(10,((double )(meausuredPower - rssi) / (10 * envFactor)));
     _ibeacon_model_state.counter++;
 }
